+++ source/main/utils/logging.js
@@ -1,9 +1,30 @@
 // @flow
 import log from 'electron-log';
+import { environment } from '../environment';
+import { formatContext } from '../../common/utils/logging';
+import type { FormatMessageContextParams } from '../../common/types/logging.types';
 
-const prefixProcessType = (str: string) => '[main] ' + str;
+const appName = 'daedalus';
+const electronProcess = 'ipcMain';
+const { network, os, platformVersion, version } = environment;
 
-const logToLevel = (level) => (message: string) => log[level](prefixProcessType(message));
+const messageContext: FormatMessageContextParams = {
+  appName,
+  electronProcess,
+  network,
+  level: '',
+};
+
+const environmentData = {
+  network,
+  os,
+  platformVersion,
+  version,
+};
+
+const logToLevel = (level: string) => (message: string, data: ?Object) => (
+  log[level](formatContext({ ...messageContext, level }), { message, data, environmentData })
+);
 
 export const Logger = {
   debug: logToLevel('debug'),